Fix some pages not showing properly

+++ app/controllers/pages_controller.rb
@@ -27,7 +27,7 @@ class PagesController < ApplicationController
 
   def show_positivity
     @users = User.all.select { |u| u.positivity_score }
-    @users = @users.sort_by { |u| u.positivity_score.to_f }.reverse
+    @users = @users.sort_by { |u| u.positivity_score.to_i }.reverse
   end
 
   def callback
